% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunction.R
\name{as.data.table.DictionaryLoss}
\alias{as.data.table.DictionaryLoss}
\title{Convert dictionary to data table}
\usage{
\method{as.data.table}{DictionaryLoss}(x, ..., objects = FALSE)
}
\arguments{
\item{x}{(`DictionaryLoss`)\cr
The dictionary to convert.}

\item{...}{(`any`)\cr
Additional arguments.}

\item{objects}{(`logical(1)`)\cr
Whether to include the objects in the result.}
}
\description{
Converts a loss function dictionary to a data table.
}
