% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_size_control.R
\name{assert_step_size_control}
\alias{assert_step_size_control}
\title{Assertion for the main signature of a \code{step_size_controlXX} function.}
\usage{
assert_step_size_control(x, u, obj, opt)
}
\arguments{
\item{x}{(\code{numeric()}) The "old" point.}

\item{u}{(\code{numeric()}) The update added to \code{x} without a step size control.}

\item{obj}{(\code{Objective}) The usd objective object.}

\item{opt}{(\code{Optimizer}) The optimizer object from which the function is called.}
}
\description{
Assertion for the main signature of a \code{step_size_controlXX} function.
}
