% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizer.R
\name{colSampler}
\alias{colSampler}
\title{Randomly generate colors}
\usage{
colSampler(alpha = NULL)
}
\arguments{
\item{alpha}{(`numeric(1)`) The alpha value. If `!is.null` the used prefix is 'rgba' instead of 'rgb'.}
}
\value{
A character of length one containing the RGB color.
}
\description{
Helper function to generate RGB colors.
}
