% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunction.R
\docType{data}
\name{dict_loss}
\alias{dict_loss}
\title{Dictionary of Loss Functions}
\format{
An object of class \code{DictionaryLoss} (inherits from \code{Dictionary}, \code{R6}) of length 12.
}
\usage{
dict_loss
}
\description{
Dictionary of loss functions.
}
\keyword{datasets}
