% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\docType{data}
\name{dict_objective}
\alias{dict_objective}
\title{Objective dictionary}
\format{
A \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} where each entry is an \link{Objective} instance.
}
\usage{
dict_objective
}
\description{
Lookup table for reusable \link{Objective} definitions. The dictionary ships with
vistool and comes pre-populated with regularized regression objectives as
well as a selection of benchmark test functions when the optional
\code{TestFunctions} package is available. New objectives can be registered at
runtime via \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} methods such as \verb{$add()}.
}
\examples{
dict_objective$get("TF_branin")

}
\seealso{
\code{\link[=obj]{obj()}} for convenient retrieval.
}
\keyword{datasets}
