% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizerLossFunction.R
\docType{data}
\name{dictionary_loss}
\alias{dictionary_loss}
\title{Dictionary of Loss Functions}
\format{
An object of class \code{DictionaryLoss} (inherits from \code{Dictionary}, \code{R6}) of length 13.
}
\usage{
dictionary_loss
}
\description{
Dictionary of loss functions.
}
\keyword{datasets}
