% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_management.R
\docType{data}
\name{.vistool_colors}
\alias{.vistool_colors}
\title{Color Management for vistool}
\format{
An object of class \code{list} of length 4.
}
\usage{
.vistool_colors
}
\description{
Internal color management system to ensure consistent colors across
ggplot2 and plotly visualizations.
}
\keyword{internal}
