% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_management.R
\name{get_continuous_colorscale}
\alias{get_continuous_colorscale}
\title{Get continuous colorscale for surface plots}
\usage{
get_continuous_colorscale(palette = "viridis")
}
\arguments{
\item{palette}{(`character(1)`)\cr
Name of the color palette. One of "viridis", "plasma", "grayscale".}
}
\value{
A list defining the colorscale for plotly.
}
\description{
Returns a continuous colorscale definition suitable for plotly surface plots
based on the specified color palette.
}
