% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{ggplot_add.vistool_theme}
\alias{ggplot_add.vistool_theme}
\title{Add vistool theme to ggplot2}
\usage{
\method{ggplot_add}{vistool_theme}(object, plot, ...)
}
\arguments{
\item{object}{A vistool_theme object.}

\item{plot}{A ggplot object.}

\item{...}{Additional arguments (unused).}
}
\description{
Add vistool theme to ggplot2
}
\keyword{internal}
