% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunction.R
\name{lss}
\alias{lss}
\title{Retrieve Loss Function}
\usage{
lss(.key, ...)
}
\arguments{
\item{.key}{(`character(1)`)\cr
Key passed to the respective [dictionary][mlr3misc::Dictionary] to retrieve the object.}

\item{...}{(named `list()`)\cr
Named arguments passed to the constructor, or to be set as public field.
See [mlr3misc::dictionary_sugar_get()] for more details.}
}
\description{
Retrieve a loss function from the dictionary.
}
