% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{merge_optim_archives}
\alias{merge_optim_archives}
\title{Merge optimization archives}
\usage{
merge_optim_archives(...)
}
\arguments{
\item{...}{(\code{any})\cr
Optimization objects.}
}
\description{
Merge archives.
}
