% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{mergeOptimArchives}
\alias{mergeOptimArchives}
\title{Merge optimization archives}
\usage{
mergeOptimArchives(...)
}
\arguments{
\item{...}{(`any`)\cr
Optimization objects.}
}
\description{
Merge archives.
}
