% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\name{obj}
\alias{obj}
\title{Retrieve objective functions}
\usage{
obj(.key, ...)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve the object.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, or to be set as public field.
See \code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_get()}} for more details.}
}
\description{
Retrieve an objective function from the dictionary.
}
