% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{Optimizer}
\alias{Optimizer}
\title{Optimizer class}
\description{
This class defines the optimization technique.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{(`character(1)`)\cr
The id of the object.}

\item{\code{print_trace}}{(`logical(1)` Indicator whether to print the status of `$optimize()`.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{lr}}{(`numeric(1`) Step size of the algorithm.}

\item{\code{archive}}{(`data.table()`) Archive of all calls to `$eval_store`.}

\item{\code{objective}}{(`Objective`) The objective function.}

\item{\code{x}}{(`numeric()`) The numerical input vector used as starting point by `$optimize()`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Optimizer-new}{\code{Optimizer$new()}}
\item \href{#method-Optimizer-prepare_update_for_archive}{\code{Optimizer$prepare_update_for_archive()}}
\item \href{#method-Optimizer-update_archive}{\code{Optimizer$update_archive()}}
\item \href{#method-Optimizer-set_x}{\code{Optimizer$set_x()}}
\item \href{#method-Optimizer-clone}{\code{Optimizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$new(objective, x_start, id = NULL, print_trace = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(`Objective`)\cr
The objective to optimize.}

\item{\code{x_start}}{(`numeric()`)\cr
Start value of the optimization. Note, after
the first call of `$optimize()` the last value is used to continue
optimization. Get this value with `$x`.}

\item{\code{id}}{(`character(1)`)\cr
Id of the object.}

\item{\code{print_trace}}{(`logical(1)`)\cr
Indicator whether to print the status of `$optimize()`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-prepare_update_for_archive"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-prepare_update_for_archive}{}}}
\subsection{Method \code{prepare_update_for_archive()}}{
Prepare updates for adding them to the archive.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$prepare_update_for_archive(
  x_out,
  x_in,
  update,
  fval_out,
  fval_in,
  lr,
  step_size,
  objective,
  step,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_out}}{(`numeric()`) The new proposed point by the optimizer.}

\item{\code{x_in}}{(`numeric()`) The old input value which is updated to `x_out`.}

\item{\code{update}}{(`numeric()`) The update from `x_in` to `x_out`.}

\item{\code{fval_out}}{(`numeric(1)`) The objective value `objetive$eval(x_out)`.}

\item{\code{fval_in}}{(`numeric(1)`) The objective value `objetive$eval(x_in)`.}

\item{\code{lr}}{(`numeric(1)`) The learning rate used to multiply `update` with.}

\item{\code{step_size}}{(`numeric(1)`) The step_size used to multiply `lr * update` with.}

\item{\code{objective}}{(`Objective`) The objective used by `$optimize()`.}

\item{\code{step}}{(`integer(1)`) The step or iteration.}

\item{\code{...}}{Additional objects added to the archive (e.g. `momentum`).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`data.table()` of the input arguments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-update_archive"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-update_archive}{}}}
\subsection{Method \code{update_archive()}}{
Add points to the archive.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$update_archive(ain)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ain}}{`data.table()` with names "x_out", "x_in", "update", "fval_out",
"fval_in", "lr", "objective", and "step".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-set_x"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-set_x}{}}}
\subsection{Method \code{set_x()}}{
Set the current input vector used as start point of `$optimize()`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$set_x(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(`numeric()`) Input vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
