% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{OptimizerGD}
\alias{OptimizerGD}
\title{Gradient descent optimizer}
\description{
Gradient descent optimizer

Gradient descent optimizer
}
\details{
This class defines gradient descent
}
\section{Super classes}{
\code{\link[vistool:Optimizer]{vistool::Optimizer}} -> \code{\link[vistool:OptimizerMomentum]{vistool::OptimizerMomentum}} -> \code{OptimizerGD}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{momentum}}{(`numeric(1)`) Momentum of the algorithm.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizerGD$new()}}
\item \href{#method-clone}{\code{OptimizerGD$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="prepareUpdateForArchive">}\href{../../vistool/html/Optimizer.html#method-prepareUpdateForArchive}{\code{vistool::Optimizer$prepareUpdateForArchive()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="setX">}\href{../../vistool/html/Optimizer.html#method-setX}{\code{vistool::Optimizer$setX()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="updateArchive">}\href{../../vistool/html/Optimizer.html#method-updateArchive}{\code{vistool::Optimizer$updateArchive()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vistool" data-topic="OptimizerMomentum" data-id="optimize">}\href{../../vistool/html/OptimizerMomentum.html#method-optimize}{\code{vistool::OptimizerMomentum$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vistool" data-topic="OptimizerMomentum" data-id="update">}\href{../../vistool/html/OptimizerMomentum.html#method-update}{\code{vistool::OptimizerMomentum$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerGD$new(
  objective,
  x_start,
  lr = 0.01,
  id = "Gradient Descent",
  print_trace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(`Objective`) The objective to optimize.}

\item{\code{x_start}}{(`numeric()`) Start value of the optimization. Note, after
the first call of `$optimize()` the last value is used to continue
optimization. Get this value with `$x`.}

\item{\code{lr}}{(`numeric(1)`) Step size with which the update is multiplied.}

\item{\code{id}}{(`character(1)`) Id of the object.}

\item{\code{print_trace}}{(`logical(1)`) Indicator whether to print the status of `$optimize()`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerGD$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
