% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{OptimizerMomentum}
\alias{OptimizerMomentum}
\title{Momentum optimizer}
\description{
Momentum optimizer

Momentum optimizer
}
\details{
This class defines momentum.
}
\section{Super class}{
\code{\link[vistool:Optimizer]{vistool::Optimizer}} -> \code{OptimizerMomentum}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{momentum}}{(`numeric(1)`) Momentum of the algorithm.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizerMomentum$new()}}
\item \href{#method-optimize}{\code{OptimizerMomentum$optimize()}}
\item \href{#method-update}{\code{OptimizerMomentum$update()}}
\item \href{#method-clone}{\code{OptimizerMomentum$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="prepareUpdateForArchive">}\href{../../vistool/html/Optimizer.html#method-prepareUpdateForArchive}{\code{vistool::Optimizer$prepareUpdateForArchive()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="setX">}\href{../../vistool/html/Optimizer.html#method-setX}{\code{vistool::Optimizer$setX()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="updateArchive">}\href{../../vistool/html/Optimizer.html#method-updateArchive}{\code{vistool::Optimizer$updateArchive()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$new(
  objective,
  x_start,
  lr = 0.01,
  momentum = 0.9,
  id = "Momentum",
  print_trace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(`Objective`) The objective to optimize.}

\item{\code{x_start}}{(`numeric()`) Start value of the optimization. Note, after
the first call of `$optimize()` the last value is used to continue
optimization. Get this value with `$x`.}

\item{\code{lr}}{(`numeric(1)`) Step size with which the update is multiplied.}

\item{\code{momentum}}{(`numeric(1)`) Momentum value.}

\item{\code{id}}{(`character(1)`) Id of the object.}

\item{\code{print_trace}}{(`logical(1)`) Indicator whether to print the status of `$optimize()`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimize `steps` iteration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$optimize(
  steps = 1L,
  stepSizeControl = function(x, u, obj, opt) return(1),
  minimize = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{steps}}{(`integer(1)`) Number of steps/iterations.}

\item{\code{stepSizeControl}}{(`function()`) A function with arguments `x` (the old input value),
`u` (the upate generated by `$update()`), `obj` (the objective object), and `opt` the optimizer as `self`.
Default simply returns 1 to include the update as is. Note that the update `u` is calculated based on the
formula with learning rate, e.g., for GD x_new = x_old + lr * g, with update u = lr * g. Step size control takes
x_old and u (with included learning rate) and calculates a value that is used to extend or compress u for the
final update.}

\item{\code{minimize}}{(`logical(1)`) Indicator to whether minimize or optimize the objective. The default (`NULL`)
uses the option defined in `objective$minimize`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Calculate the update for `x`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$update(lr, mom)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lr}}{(`numeric(1)`) The learning rate.}

\item{\code{mom}}{(`numeric(1)`) The momentum.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMomentum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
