% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{OptimizerNR}
\alias{OptimizerNR}
\title{Newton-Raphson optimizer}
\description{
Implements Newton-Raphson updates by combining gradients and Hessians of the
objective with optional Armijo backtracking. Falls back to the gradient
direction when the Hessian is singular or indefinite.
}
\section{Super class}{
\code{\link[vistool:Optimizer]{vistool::Optimizer}} -> \code{OptimizerNR}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizerNR-new}{\code{OptimizerNR$new()}}
\item \href{#method-OptimizerNR-optimize}{\code{OptimizerNR$optimize()}}
\item \href{#method-OptimizerNR-step_size}{\code{OptimizerNR$step_size()}}
\item \href{#method-OptimizerNR-gamma}{\code{OptimizerNR$gamma()}}
\item \href{#method-OptimizerNR-tau}{\code{OptimizerNR$tau()}}
\item \href{#method-OptimizerNR-clone}{\code{OptimizerNR$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="prepare_update_for_archive"><a href='../../vistool/html/Optimizer.html#method-Optimizer-prepare_update_for_archive'><code>vistool::Optimizer$prepare_update_for_archive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="set_x"><a href='../../vistool/html/Optimizer.html#method-Optimizer-set_x'><code>vistool::Optimizer$set_x()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="update_archive"><a href='../../vistool/html/Optimizer.html#method-Optimizer-update_archive'><code>vistool::Optimizer$update_archive()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNR-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNR-new}{}}}
\subsection{Method \code{new()}}{
Creates a new Newton-Raphson optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNR$new(
  objective,
  x_start,
  step_size = 1,
  gamma = 0.99,
  tau = 0.5,
  max_backtracks = 20L,
  fallback = c("gradient", "stop"),
  id = "Newton-Raphson",
  print_trace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(\code{Objective})\cr
The objective to optimize.}

\item{\code{objective}}{(\code{Objective})\cr
The objective to optimize.}

\item{\code{x_start}}{(\code{numeric()})\cr
Start value of the optimization. Note, after
the first call of \verb{$optimize()} the last value is used to continue
optimization. Get this value with \verb{$x}.}

\item{\code{x_start}}{(\code{numeric()})\cr
Start value of the optimization. Note, after
the first call of \verb{$optimize()} the last value is used to continue
optimization. Get this value with \verb{$x}.}

\item{\code{step_size}}{(\code{numeric(1)})
Initial step size used as the starting value for line search. Default \code{1}.}

\item{\code{gamma}}{(\code{numeric(1)})
Armijo slope parameter in \verb{(0, 1)}.}

\item{\code{tau}}{(\code{numeric(1)})
Multiplicative backtracking factor in \verb{(0, 1)}.}

\item{\code{max_backtracks}}{(\code{integer(1)})
Maximum number of Armijo backtracking attempts.}

\item{\code{fallback}}{(\code{character(1)})
Strategy when the Hessian cannot be inverted. One of "gradient" or "stop".}

\item{\code{id}}{(\code{character(1)})\cr
Id of the object.}

\item{\code{id}}{(\code{character(1)})\cr
Id of the object.}

\item{\code{print_trace}}{(\code{logical(1)})\cr
Indicator whether to print the status of \verb{$optimize()}.}

\item{\code{print_trace}}{(\code{logical(1)})\cr
Indicator whether to print the status of \verb{$optimize()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNR-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNR-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimize the associated objective for a number of steps.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNR$optimize(steps = 1L, step_size_control = NULL, minimize = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{steps}}{(\code{integer(1)})\cr
Number of steps/iterations.}

\item{\code{step_size_control}}{(\verb{function()})
Optional callback with signature \verb{(x, u, obj, opt)} returning a scalar
step size. Defaults to Armijo backtracking that uses the configured
\code{gamma}, \code{tau}, and \code{max_backtracks} parameters.}

\item{\code{minimize}}{(\code{logical(1)})
Whether to minimize the objective. Defaults to
\code{self$objective$minimize}. Maximization is not supported and triggers an error.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNR-step_size"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNR-step_size}{}}}
\subsection{Method \code{step_size()}}{
Access or update the base step size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNR$step_size(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{numeric(1)})
Replacement value. Omit to retrieve the current setting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNR-gamma"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNR-gamma}{}}}
\subsection{Method \code{gamma()}}{
Access or update the Armijo slope parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNR$gamma(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{numeric(1)})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNR-tau"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNR-tau}{}}}
\subsection{Method \code{tau()}}{
Access or update the Armijo shrink factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNR$tau(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{numeric(1)})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
