% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_size_control.R
\name{step_size_control_decay_linear}
\alias{step_size_control_decay_linear}
\title{Conduct linear decay to adjust the update.
See https://neptune.ai/blog/how-to-choose-a-learning-rate-scheduler}
\usage{
step_size_control_decay_linear(iter_zero = 100L)
}
\arguments{
\item{iter_zero}{(\code{integer(1)}) The iteration at which the update is shrinked to zero.}
}
\value{
The step size as number.
}
\description{
Conduct linear decay to adjust the update.
See https://neptune.ai/blog/how-to-choose-a-learning-rate-scheduler
}
