% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_size_control.R
\name{step_size_control_decay_steps}
\alias{step_size_control_decay_steps}
\title{Conduct a step-wise decay to adjust the update.
See https://neptune.ai/blog/how-to-choose-a-learning-rate-scheduler}
\usage{
step_size_control_decay_steps(drop_rate = 0.1, every_iter = 10)
}
\arguments{
\item{drop_rate}{(`numeric(1)`) The rate indicating how much the learning rate is reduced
after each `every_iter`.}

\item{every_iter}{(`integer(1)`) Number indicates after how many iterations the
learning rate is reduced by `drop_rate`.}
}
\value{
The step size as number.
}
\description{
Conduct a step-wise decay to adjust the update.
See https://neptune.ai/blog/how-to-choose-a-learning-rate-scheduler
}
