% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_size_control.R
\name{step_size_control_line_search}
\alias{step_size_control_line_search}
\title{Conduct line search in each iteration to adjust the update.}
\usage{
step_size_control_line_search(lower = 0, upper = 10)
}
\arguments{
\item{lower}{(\code{numeric(1)}) The lower bound for the step_size.}

\item{upper}{(\code{numeric(1)}) The upper bound for the step_size.}
}
\value{
The step size as number.
}
\description{
Conduct line search in each iteration to adjust the update.
}
