% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_size_control.R
\name{stepSizeControlDecayExp}
\alias{stepSizeControlDecayExp}
\title{Conduct exponential decay to adjust the update.
See https://neptune.ai/blog/how-to-choose-a-learning-rate-scheduler}
\usage{
stepSizeControlDecayExp(decay = 0.01)
}
\arguments{
\item{decay}{(`numeric(1)`) The decay parameter indicating how fast the step size is reduced.}
}
\value{
The step size as number.
}
\description{
Conduct exponential decay to adjust the update.
See https://neptune.ai/blog/how-to-choose-a-learning-rate-scheduler
}
