% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_size_control.R
\name{stepSizeControlLineSearch}
\alias{stepSizeControlLineSearch}
\title{Conduct line search in each iteration to adjust the update.}
\usage{
stepSizeControlLineSearch(lower = 0, upper = 10)
}
\arguments{
\item{lower}{(`numeric(1)`) The lower bound for the step_size.}

\item{upper}{(`numeric(1)`) The upper bound for the step_size.}
}
\value{
The step size as number.
}
\description{
Conduct line search in each iteration to adjust the update.
}
