% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\docType{data}
\name{tfun_dict}
\alias{tfun_dict}
\title{Dictionary for test functions}
\format{
An object of class \code{DictionaryObjective} (inherits from \code{Dictionary}, \code{R6}) of length 13.
}
\usage{
tfun_dict
}
\description{
Dictionary for test functions
}
\examples{
tfun_dict$get("TF_branin")
}
\keyword{datasets}
