% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{vistool_palette_theme}
\alias{vistool_palette_theme}
\title{Build ggplot2 palette theme elements}
\usage{
vistool_palette_theme(palette)
}
\description{
Build ggplot2 palette theme elements
}
\keyword{internal}
