% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer1D.R
\name{Visualizer1D}
\alias{Visualizer1D}
\title{Visualize Base Class}
\description{
This class is used to create 1D visualizations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x}}{(`vector()`)\cr
x-values.}

\item{\code{y}}{(`vector()`)\cr
y-values}

\item{\code{plot_lab}}{(character(1)\cr
Label of the plot.}

\item{\code{x_lab}}{(character(1)\cr
Label of the x axis.}

\item{\code{y_lab}}{(character(1)\cr
Label of the y axis.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer1D-new}{\code{Visualizer1D$new()}}
\item \href{#method-Visualizer1D-init_layer_lines}{\code{Visualizer1D$init_layer_lines()}}
\item \href{#method-Visualizer1D-setLayout}{\code{Visualizer1D$setLayout()}}
\item \href{#method-Visualizer1D-plot}{\code{Visualizer1D$plot()}}
\item \href{#method-Visualizer1D-save}{\code{Visualizer1D$save()}}
\item \href{#method-Visualizer1D-clone}{\code{Visualizer1D$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1D-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1D-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1D$new(x, y, plot_lab = NULL, x_lab = "x", y_lab = "y")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(`numeric()`)\cr
x-values.}

\item{\code{y}}{(`numeric()`)\cr
y-values.}

\item{\code{plot_lab}}{(character(1)\cr
Label of the plot.}

\item{\code{x_lab}}{(character(1)\cr
Label of the x axis.}

\item{\code{y_lab}}{(character(1)\cr
Label of the y axis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1D-init_layer_lines"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1D-init_layer_lines}{}}}
\subsection{Method \code{init_layer_lines()}}{
Initialize the plot with a line plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1D$init_layer_lines(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(`any`)\cr
Further arguments passed to `add_trace(...)`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1D-setLayout"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1D-setLayout}{}}}
\subsection{Method \code{setLayout()}}{
Set the layout of the plotly plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1D$setLayout(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(`any`)\cr
Layout options directly passed to `layout(...)`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1D-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1D-plot}{}}}
\subsection{Method \code{plot()}}{
Return the plot and hence plot it or do further processing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1D$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1D-save"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1D-save}{}}}
\subsection{Method \code{save()}}{
Save the plot by using plotlys `orca()` function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1D$save(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(`any`)\cr
Further arguments passed to `orca()`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1D-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1D-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
