% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer1DObjective.R
\name{Visualizer1DObjective}
\alias{Visualizer1DObjective}
\title{Visualize Objective}
\description{
This class is used to create visualizations of optimization traces.
}
\section{Super class}{
\code{\link[vistool:Visualizer1D]{vistool::Visualizer1D}} -> \code{Visualizer1DObjective}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objective}}{(`Objective`)\cr
The objective which was optimized.
This object is used to generate the surface/contour lines.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer1DObjective-new}{\code{Visualizer1DObjective$new()}}
\item \href{#method-Visualizer1DObjective-add_optimization_trace}{\code{Visualizer1DObjective$add_optimization_trace()}}
\item \href{#method-Visualizer1DObjective-clone}{\code{Visualizer1DObjective$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer1D" data-id="init_layer_lines"><a href='../../vistool/html/Visualizer1D.html#method-Visualizer1D-init_layer_lines'><code>vistool::Visualizer1D$init_layer_lines()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer1D" data-id="plot"><a href='../../vistool/html/Visualizer1D.html#method-Visualizer1D-plot'><code>vistool::Visualizer1D$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer1D" data-id="save"><a href='../../vistool/html/Visualizer1D.html#method-Visualizer1D-save'><code>vistool::Visualizer1D$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer1D" data-id="setLayout"><a href='../../vistool/html/Visualizer1D.html#method-Visualizer1D-setLayout'><code>vistool::Visualizer1D$setLayout()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1DObjective-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1DObjective-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1DObjective$new(
  objective,
  x_limits = NULL,
  padding = 0,
  n_points = 100L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(`Objective`)\cr
The objective which was optimized.
This object is used to generate the surface/contour lines.}

\item{\code{x_limits}}{(`numeric(2)`)\cr
The limits of the x axis.}

\item{\code{padding}}{(`numeric(1)`)\cr
A margin that is added to x1limits and x2limits.
The x1 margin is calculated by `max(x1lmits) - min(x1limits) * padding`.}

\item{\code{n_points}}{(`integer(1)`)\cr
The number of generated point per dimension.
Note that a grid of `npoints^2` values is generated and evaluated by `objective$eval(x)` to plot the surface.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1DObjective-add_optimization_trace"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1DObjective-add_optimization_trace}{}}}
\subsection{Method \code{add_optimization_trace()}}{
Add optimization trace to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1DObjective$add_optimization_trace(optimizer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{(`Optimizer`)\cr
The optimizer to add to the plot.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1DObjective-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1DObjective-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1DObjective$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
