% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer3DModel.R
\name{Visualizer3DModel}
\alias{Visualizer3DModel}
\title{Visualize Model}
\description{
This class is used to create 3D visualizations of learners and tasks.
}
\section{Super classes}{
\code{\link[vistool:Visualizer]{vistool::Visualizer}} -> \code{\link[vistool:Visualizer3D]{vistool::Visualizer3D}} -> \code{Visualizer3DModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{(`mlr3::Task`)\cr
Task used to train the model.}

\item{\code{learner}}{(`mlr3::Learner`)\cr
Learner used to train the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer3DModel-new}{\code{Visualizer3DModel$new()}}
\item \href{#method-Visualizer3DModel-add_training_data}{\code{Visualizer3DModel$add_training_data()}}
\item \href{#method-Visualizer3DModel-add_decision_boundary}{\code{Visualizer3DModel$add_decision_boundary()}}
\item \href{#method-Visualizer3DModel-clone}{\code{Visualizer3DModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="save"><a href='../../vistool/html/Visualizer.html#method-Visualizer-save'><code>vistool::Visualizer$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer3D" data-id="init_layer_contour"><a href='../../vistool/html/Visualizer3D.html#method-Visualizer3D-init_layer_contour'><code>vistool::Visualizer3D$init_layer_contour()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer3D" data-id="init_layer_surface"><a href='../../vistool/html/Visualizer3D.html#method-Visualizer3D-init_layer_surface'><code>vistool::Visualizer3D$init_layer_surface()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer3D" data-id="plot"><a href='../../vistool/html/Visualizer3D.html#method-Visualizer3D-plot'><code>vistool::Visualizer3D$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer3D" data-id="set_layout"><a href='../../vistool/html/Visualizer3D.html#method-Visualizer3D-set_layout'><code>vistool::Visualizer3D$set_layout()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer3D" data-id="set_scene"><a href='../../vistool/html/Visualizer3D.html#method-Visualizer3D-set_scene'><code>vistool::Visualizer3D$set_scene()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer3DModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer3DModel-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer3DModel$new(
  task,
  learner,
  x1_limits = NULL,
  x2_limits = NULL,
  padding = 0,
  n_points = 100L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{([mlr3::Task])\cr
The task to train the model on.}

\item{\code{learner}}{(`mlr3::Learner`)\cr
The learner to train the model with.}

\item{\code{x1_limits}}{(`numeric(2)`)\cr
The x1 limits.}

\item{\code{x2_limits}}{(`numeric(2)`)\cr
The x2 limits.}

\item{\code{padding}}{(`numeric(1)`)\cr
A margin that is added to x1limits and x2limits.
The x1 margin is calculated by `max(x1lmits) - min(x1limits) * padding`.}

\item{\code{n_points}}{(`integer(1)`)\cr
The number of generated point per dimension.
Note that a grid of `npoints^2` values is generated and evaluated by `objective$eval(x)` to plot the surface.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer3DModel-add_training_data"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer3DModel-add_training_data}{}}}
\subsection{Method \code{add_training_data()}}{
Adds the training data to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer3DModel$add_training_data(size = 5, color = "grey", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{(`numeric(1)`)\cr
Size of the points.}

\item{\code{color}}{(`character(1)`)\cr
Color of the points.}

\item{\code{...}}{(`any`)\cr
Further arguments passed to `add_trace(...)`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer3DModel-add_decision_boundary"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer3DModel-add_decision_boundary}{}}}
\subsection{Method \code{add_decision_boundary()}}{
Adds the decision boundary to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer3DModel$add_decision_boundary(threshold = 0.5, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{(`numeric(1)`)\cr
Threshold for the decision boundary.}

\item{\code{...}}{(`any`)\cr
Further arguments passed to `add_trace(...)`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer3DModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer3DModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer3DModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
