% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizerLossFunction.R
\name{VisualizerLossFunction}
\alias{VisualizerLossFunction}
\title{Visualize Loss Function}
\description{
This class is used to create visualizations of loss functions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{loss_function}}{[LossFunction]\cr
Loss function.}

\item{\code{loss}}{(`numeric()`)\cr
Loss values.}

\item{\code{y_pred}}{`numeric()`\cr
Predicted values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VisualizerLossFunction-new}{\code{VisualizerLossFunction$new()}}
\item \href{#method-VisualizerLossFunction-init_layer_lines}{\code{VisualizerLossFunction$init_layer_lines()}}
\item \href{#method-VisualizerLossFunction-set_layout}{\code{VisualizerLossFunction$set_layout()}}
\item \href{#method-VisualizerLossFunction-plot}{\code{VisualizerLossFunction$plot()}}
\item \href{#method-VisualizerLossFunction-save}{\code{VisualizerLossFunction$save()}}
\item \href{#method-VisualizerLossFunction-clone}{\code{VisualizerLossFunction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFunction-new"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFunction-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFunction$new(loss_function, y_pred, y_true, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loss_function}}{[LossFunction]\cr
Loss function.}

\item{\code{y_pred}}{(`numeric()`)\cr
Predicted values.}

\item{\code{y_true}}{(`numeric(1)`)\cr
True value.}

\item{\code{...}}{(`any`)\cr
Additional arguments passed to the loss function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFunction-init_layer_lines"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFunction-init_layer_lines}{}}}
\subsection{Method \code{init_layer_lines()}}{
Initialize the plot with a line plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFunction$init_layer_lines(
  width = 2,
  color = "rgb(160,82,45)",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{(`integer(1)`)\cr
Width of the line.}

\item{\code{color}}{(`character(1)`)\cr
Color of the line.}

\item{\code{...}}{(`any`)\cr
Further arguments passed to `add_trace(...)`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFunction-set_layout"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFunction-set_layout}{}}}
\subsection{Method \code{set_layout()}}{
Set the layout of the plotly plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFunction$set_layout(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(`any`)\cr
Layout options directly passed to `layout(...)`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFunction-plot"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFunction-plot}{}}}
\subsection{Method \code{plot()}}{
Return the plot and hence plot it or do further processing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFunction$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFunction-save"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFunction-save}{}}}
\subsection{Method \code{save()}}{
Save the plot by using plotlys `orca()` function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFunction$save(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(`any`)\cr
Further arguments passed to `orca()`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFunction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFunction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFunction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
