% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizerLossFuns.R
\name{VisualizerLossFuns}
\alias{VisualizerLossFuns}
\title{Visualizer for Losses}
\value{
Returns self invisibly.
}
\description{
Visualize one or multiple loss functions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{losses}}{(`list`)\cr
List of LossFunction objects.}

\item{\code{task_type}}{(`character(1)`)\cr
Task type (regr or classif).}

\item{\code{lab_x}}{(`character(1)`)\cr
Label for x-axis.}

\item{\code{lab_y}}{(`character(1)`)\cr
Label for y-axis.}

\item{\code{title}}{(`character(1)`)\cr
Title of the plot.}

\item{\code{x_range}}{(`numeric(2)`)\cr
Range for x-axis values.}

\item{\code{line_width}}{(`numeric()`)\cr
Line widths for different loss functions.}

\item{\code{line_col}}{(`character()`)\cr
Line colors for different loss functions.}

\item{\code{line_type}}{(`character()`)\cr
Line types for different loss functions.}

\item{\code{legend_title}}{(`element`)\cr
Legend title element.}

\item{\code{y_pred}}{(`numeric()`)\cr
Predicted values.}

\item{\code{y_true}}{(`numeric()`)\cr
True values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VisualizerLossFuns-new}{\code{VisualizerLossFuns$new()}}
\item \href{#method-VisualizerLossFuns-init_layer_lines}{\code{VisualizerLossFuns$init_layer_lines()}}
\item \href{#method-VisualizerLossFuns-plot}{\code{VisualizerLossFuns$plot()}}
\item \href{#method-VisualizerLossFuns-clone}{\code{VisualizerLossFuns$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFuns-new"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFuns-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFuns$new(losses, y_pred = NULL, y_true = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{losses}}{(`list`)\cr
List of LossFunction objects.}

\item{\code{y_pred}}{(`numeric()`)\cr
Predicted values. Optional.}

\item{\code{y_true}}{(`numeric()`)\cr
True values. Optional.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFuns-init_layer_lines"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFuns-init_layer_lines}{}}}
\subsection{Method \code{init_layer_lines()}}{
Initialize line layer for the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFuns$init_layer_lines(color = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color}}{(`character()`)\cr
Colors to use for the lines. If NULL, default colors will be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFuns-plot"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFuns-plot}{}}}
\subsection{Method \code{plot()}}{
Create and return the ggplot2 plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFuns$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 object showing the loss functions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerLossFuns-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerLossFuns-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerLossFuns$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
