% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizerObj.R
\name{VisualizerObj}
\alias{VisualizerObj}
\title{Visualize objective (unified 1D/2D)}
\value{
Returns self invisibly.
}
\description{
This class provides a unified interface for visualizing objective functions
and optimization traces on both 1D and 2D objectives. It automatically detects
the dimensionality and creates appropriate visualizations using ggplot2.
}
\section{Super class}{
\code{\link[vistool:Visualizer]{vistool::Visualizer}} -> \code{VisualizerObj}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objective}}{(\code{Objective})\cr
The objective function.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VisualizerObj-new}{\code{VisualizerObj$new()}}
\item \href{#method-VisualizerObj-add_optimization_trace}{\code{VisualizerObj$add_optimization_trace()}}
\item \href{#method-VisualizerObj-plot}{\code{VisualizerObj$plot()}}
\item \href{#method-VisualizerObj-clone}{\code{VisualizerObj$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="add_annotation"><a href='../../vistool/html/Visualizer.html#method-Visualizer-add_annotation'><code>vistool::Visualizer$add_annotation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="add_contours"><a href='../../vistool/html/Visualizer.html#method-Visualizer-add_contours'><code>vistool::Visualizer$add_contours()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="add_points"><a href='../../vistool/html/Visualizer.html#method-Visualizer-add_points'><code>vistool::Visualizer$add_points()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="resolve_layer_colors"><a href='../../vistool/html/Visualizer.html#method-Visualizer-resolve_layer_colors'><code>vistool::Visualizer$resolve_layer_colors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="save"><a href='../../vistool/html/Visualizer.html#method-Visualizer-save'><code>vistool::Visualizer$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="set_theme"><a href='../../vistool/html/Visualizer.html#method-Visualizer-set_theme'><code>vistool::Visualizer$set_theme()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="theme"><a href='../../vistool/html/Visualizer.html#method-Visualizer-theme'><code>vistool::Visualizer$theme()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerObj-new"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerObj-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerObj$new(
  objective,
  x1_limits = NULL,
  x2_limits = NULL,
  padding = 0,
  n_points = 100L,
  type = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(\code{Objective})\cr
The objective to optimize.}

\item{\code{x1_limits}}{(\code{numeric(2)})\cr
Limits for the first dimension. For 1D objectives, this controls the x-axis range.
If NULL, will be determined from objective bounds.}

\item{\code{x2_limits}}{(\code{numeric(2)})\cr
Limits for the second dimension (2D objectives only). Ignored for 1D objectives.
If NULL, will be determined from objective bounds.}

\item{\code{padding}}{(\code{numeric(1)})\cr
A margin that is added to x1limits and x2limits.
The x1 margin is calculated by \code{max(x1lmits) - min(x1limits) * padding}.}

\item{\code{padding}}{(\code{numeric(1)})\cr
A margin that is added to x1limits and x2limits.
The x1 margin is calculated by \code{max(x1lmits) - min(x1limits) * padding}.}

\item{\code{n_points}}{(\code{integer(1)})\cr
The number of generated point per dimension.
Note that a grid of \code{npoints^2} values is generated and evaluated by \code{objective$eval(x)} to plot the surface.}

\item{\code{n_points}}{(\code{integer(1)})\cr
The number of generated point per dimension.
Note that a grid of \code{npoints^2} values is generated and evaluated by \code{objective$eval(x)} to plot the surface.}

\item{\code{type}}{(\code{character(1)})\cr
Optional visualization type ("1d" or "2d") to override automatic dimension detection.
Useful when objective dimensions are unknown but visualization type is explicitly specified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerObj-add_optimization_trace"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerObj-add_optimization_trace}{}}}
\subsection{Method \code{add_optimization_trace()}}{
Add optimization trace to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerObj$add_optimization_trace(
  optimizer,
  line_color = NULL,
  line_width = NULL,
  line_type = "solid",
  npoints = NULL,
  npmax = NULL,
  name = NULL,
  add_marker_at = 1,
  marker_size = 3,
  marker_shape = 16,
  marker_color = NULL,
  show_start_end = TRUE,
  alpha = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{(\code{Optimizer})\cr
The optimizer to add to the plot. Must have been run and contain archive data.}

\item{\code{line_color}}{(\code{character(1)})\cr
Color of the optimization trace line. If NULL, uses automatic color.}

\item{\code{line_width}}{(\code{numeric(1)})\cr
Width of the trace line. If NULL, uses theme$line_width.}

\item{\code{line_type}}{(\code{character(1)})\cr
Type of the trace line. One of "solid", "dashed", "dotted". Default is "solid".}

\item{\code{npoints}}{(\code{integer(1)})\cr
Number of points to show from the trace. If NULL, shows all points.}

\item{\code{npmax}}{(\code{integer(1)})\cr
Maximum number of points to show. If NULL, no limit.}

\item{\code{name}}{(\code{character(1)})\cr
Name for the trace (used in legends). If NULL, uses optimizer ID.}

\item{\code{add_marker_at}}{(\code{integer()})\cr
Iteration numbers where to add markers. For 2D plots only.}

\item{\code{marker_size}}{(\code{numeric(1)})\cr
Size of markers. Default is 3.}

\item{\code{marker_shape}}{(\code{numeric(1)} or \code{character(1)})\cr
Shape of markers. Default is 16 (filled circle).}

\item{\code{marker_color}}{(\code{character(1)})\cr
Color of markers. If NULL, uses line color.}

\item{\code{show_start_end}}{(\code{logical(1)})\cr
Whether to highlight start and end points. Default is TRUE for 2D plots.}

\item{\code{alpha}}{(\code{numeric(1)})\cr
Alpha transparency. If NULL, uses theme$alpha.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerObj-plot"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerObj-plot}{}}}
\subsection{Method \code{plot()}}{
Create and return the ggplot2 plot with model-specific layers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerObj$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the parent plot method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VisualizerObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VisualizerObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VisualizerObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
