% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\docType{data}
\name{dict_objective}
\alias{dict_objective}
\title{Dictionary for test functions}
\format{
An object of class \code{DictionaryObjective} (inherits from \code{Dictionary}, \code{R6}) of length 13.
}
\usage{
dict_objective
}
\description{
Dictionary for test functions
}
\examples{
dict_objective$get("TF_branin")
}
\keyword{datasets}
