% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunction.R
\name{lss}
\alias{lss}
\title{Retrieve loss function}
\usage{
lss(.key, ...)
}
\arguments{
\item{.key}{(`character(1)`)\cr
Key passed to the respective [dictionary][mlr3misc::Dictionary] to retrieve the object.}

\item{...}{(`any`)\cr
Additional arguments.}
}
\description{
Retrieve a loss function from the dictionary. If additional parameters are provided,
they will be used to customize the loss function (e.g., quantile for pinball loss).
}
