% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{mergeOptimArchives}
\alias{mergeOptimArchives}
\title{Merge optimization archives}
\usage{
mergeOptimArchives(...)
}
\arguments{
\item{...}{Optimization objects.}
}
\description{
Merge optimization archives
}
