% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\name{obj}
\alias{obj}
\title{Retrieve Objective Functions}
\usage{
obj(.key, ...)
}
\arguments{
\item{.key}{(`character(1)`)\cr
Key passed to the respective [dictionary][mlr3misc::Dictionary] to retrieve the object.}

\item{...}{(named `list()`)\cr
Named arguments passed to the constructor, or to be set as public field.
See [mlr3misc::dictionary_sugar_get()] for more details.}
}
\description{
Retrieve an objective function from the dictionary.
}
