% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\name{Objective}
\alias{Objective}
\title{Objective function}
\value{
Returns self invisibly.
}
\description{
This class defines the objective that is used for optimization.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
The id of the object.}

\item{\code{label}}{(\code{character(1)} The label of the objective, i.e. a.}

\item{\code{lower}}{(\code{numeric()}) The lower limits for each dimension.}

\item{\code{upper}}{(\code{numeric()}) The upper limits for each dimension.}

\item{\code{minimize}}{(\code{logical(1)}) Is the problem a minimization problem?}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{archive}}{(\code{data.table()}) Archive of all calls to \verb{$eval_store}.}

\item{\code{log_funs}}{(\code{list()}) A list containing logging functions. Each function must have argument.}

\item{\code{xdim}}{(\code{integer(1)}) Input dimension of \code{f}.}

\item{\code{transform_id}}{(\code{character(1)}) Identifier of the active transform.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Objective-new}{\code{Objective$new()}}
\item \href{#method-Objective-eval}{\code{Objective$eval()}}
\item \href{#method-Objective-eval_store}{\code{Objective$eval_store()}}
\item \href{#method-Objective-assert_x}{\code{Objective$assert_x()}}
\item \href{#method-Objective-set_transform}{\code{Objective$set_transform()}}
\item \href{#method-Objective-get_transform}{\code{Objective$get_transform()}}
\item \href{#method-Objective-grad}{\code{Objective$grad()}}
\item \href{#method-Objective-hess}{\code{Objective$hess()}}
\item \href{#method-Objective-add_log_fun}{\code{Objective$add_log_fun()}}
\item \href{#method-Objective-clear_archive}{\code{Objective$clear_archive()}}
\item \href{#method-Objective-clone}{\code{Objective$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-new"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$new(
  id,
  fun,
  label = "f",
  xdim,
  lower = NA,
  upper = NA,
  xtest = NULL,
  minimize = FALSE,
  transform = objective_transform_identity(),
  label_base = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Id of the object.}

\item{\code{fun}}{(\code{function} The objective function. The first argument must be a numerical input of length \code{xdim}.}

\item{\code{label}}{(\code{character(1)} The label of the objective, i.e. a.}

\item{\code{xdim}}{(\code{integer(1)}) The input dimension of \code{fun}. Use \code{xdim = NA} for an arbitrary input dimension.}

\item{\code{lower}}{(\code{numeric(xdim)}) The lower boundaries for inputs to \code{fun}. Must}

\item{\code{upper}}{(\code{numeric(xdim)}) The upper boundaries for inputs to \code{fun}. Must
be of length \code{xdim}.}

\item{\code{xtest}}{(\code{numeric()}) Test value for \code{fun} during initialization. If not defined,
\code{xtest = rep(0, ifelse(is.na(xdim), 2, xdim))} is used.}

\item{\code{minimize}}{(\code{logical(1)}) Is the problem a minimization problem? Default is no (\code{FALSE}).}

\item{\code{transform}}{(\code{objective_transform()}) Optional scalar transformation applied
to the objective value. Defaults to \code{\link[=objective_transform_identity]{objective_transform_identity()}}.}

\item{\code{label_base}}{(\code{character(1)}) Optional base label used when appending
transform identifiers. Defaults to \code{label}.}

\item{\code{...}}{\cr
Additional arguments passed to \code{fun}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-eval"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluate the objective function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$eval(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{numeric}) The numerical input of \code{fun}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of \code{fun(x)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-eval_store"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-eval_store}{}}}
\subsection{Method \code{eval_store()}}{
Evaluate the objective function and log into the archive. Each call logs
the input vector \code{x}, result of fun \code{fval}, the gradient \code{grad}, the norm of the gradient
\code{gnorm}, and additional logs that were added by \verb{$add_log_fun}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$eval_store(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{numeric}) The numerical input of \code{fun}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible list of logs that are added to the archive.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-assert_x"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-assert_x}{}}}
\subsection{Method \code{assert_x()}}{
Assert a numeric input if it is suitable or not.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$assert_x(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{numeric()}) Input value for \code{fun}.}

\item{\code{...}}{Additional arguments passed to \code{assertNumeric(...)}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-set_transform"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-set_transform}{}}}
\subsection{Method \code{set_transform()}}{
Update the active transformation for the objective.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$set_transform(transform, test = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transform}}{(\code{objective_transform()}) Transformation descriptor created
via \code{\link[=objective_transform]{objective_transform()}}.}

\item{\code{test}}{(\code{logical(1)}) If \code{TRUE}, validates the transform by evaluating
the objective at the stored test point.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \link{Objective} invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-get_transform"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-get_transform}{}}}
\subsection{Method \code{get_transform()}}{
Retrieve the active transformation descriptor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$get_transform()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object created by \code{\link[=objective_transform]{objective_transform()}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-grad"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-grad}{}}}
\subsection{Method \code{grad()}}{
Evaluate the gradient of the objective function at x.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$grad(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{numeric}) The numerical input of \code{fun}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-hess"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-hess}{}}}
\subsection{Method \code{hess()}}{
Evaluate the hessian of the objective function at x.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$hess(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{numeric}) The numerical input of \code{fun}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-add_log_fun"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-add_log_fun}{}}}
\subsection{Method \code{add_log_fun()}}{
Method to add custom logger to the objective.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$add_log_fun(l, label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{l}}{(\code{function}) Function that returns a single numerical value or a string.
The arguments of \code{l} must be \code{x}, \code{fval} and \code{grad}.}

\item{\code{label}}{(\code{character(1)}) The name of the logger.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-clear_archive"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-clear_archive}{}}}
\subsection{Method \code{clear_archive()}}{
Delete the archive.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$clear_archive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Objective-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Objective-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Objective$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
