% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{OptimizerNAG}
\alias{OptimizerNAG}
\title{Nesterovs momentum optimizer}
\description{
This class defines Nesterovs momentum using Nesterov accelerated gradient (NAG).
}
\section{Super class}{
\code{\link[vistool:Optimizer]{vistool::Optimizer}} -> \code{OptimizerNAG}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{momentum}}{(\code{numeric(1)}) Momentum of the algorithm.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizerNAG-new}{\code{OptimizerNAG$new()}}
\item \href{#method-OptimizerNAG-optimize}{\code{OptimizerNAG$optimize()}}
\item \href{#method-OptimizerNAG-update}{\code{OptimizerNAG$update()}}
\item \href{#method-OptimizerNAG-clone}{\code{OptimizerNAG$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="prepare_update_for_archive"><a href='../../vistool/html/Optimizer.html#method-Optimizer-prepare_update_for_archive'><code>vistool::Optimizer$prepare_update_for_archive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="set_x"><a href='../../vistool/html/Optimizer.html#method-Optimizer-set_x'><code>vistool::Optimizer$set_x()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Optimizer" data-id="update_archive"><a href='../../vistool/html/Optimizer.html#method-Optimizer-update_archive'><code>vistool::Optimizer$update_archive()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNAG-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNAG-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNAG$new(
  objective,
  x_start,
  lr = 0.01,
  momentum = 0.9,
  id = "NAG",
  print_trace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(\code{Objective})\cr
The objective to optimize.}

\item{\code{x_start}}{(\code{numeric()})\cr
Start value of the optimization. Note, after
the first call of \verb{$optimize()} the last value is used to continue
optimization. Get this value with \verb{$x}.}

\item{\code{lr}}{(\code{numeric(1)})\cr
Step size with which the update is multiplied.}

\item{\code{momentum}}{(\code{numeric(1)})\cr
Momentum value.}

\item{\code{id}}{(\code{character(1)})\cr
Id of the object.}

\item{\code{print_trace}}{(\code{logical(1)})\cr
Indicator whether to print the status of \verb{$optimize()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNAG-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNAG-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimize \code{steps} iteration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNAG$optimize(
  steps = 1L,
  step_size_control = function(x, u, obj, opt) {
     return(1)
 },
  minimize = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{steps}}{(\code{integer(1)})\cr
Number of steps/iterations.}

\item{\code{step_size_control}}{(\verb{function()}) A function with arguments \code{x} (the old input value),
\code{u} (the update generated by \verb{$update()}), and \code{obj} (the objective object).}

\item{\code{minimize}}{(\code{logical(1)}) Indicator to whether minimize or optimize the objective. The default (\code{NULL})
uses the option defined in \code{objective$minimize}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNAG-update"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNAG-update}{}}}
\subsection{Method \code{update()}}{
Calculate the update for \code{x}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNAG$update(lr, mom, minimize)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lr}}{(\code{numeric(1)}) The learning rate.}

\item{\code{mom}}{(\code{numeric(1)}) The momentum.}

\item{\code{minimize}}{(\code{logical(1)}) Indicator to whether minimize or optimize the objective (default = \code{TRUE}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerNAG-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerNAG-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerNAG$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
