% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{Visualizer}
\alias{Visualizer}
\title{Base Visualizer Class}
\description{
Base class for all visualizers. Provides a common interface for creating
and saving plots across different plotting backends (ggplot2 for 1D/2D, plotly for 3D).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer-plot}{\code{Visualizer$plot()}}
\item \href{#method-Visualizer-save}{\code{Visualizer$save()}}
\item \href{#method-Visualizer-clone}{\code{Visualizer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer-plot}{}}}
\subsection{Method \code{plot()}}{
Abstract method to be implemented by subclasses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer$plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The plot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer-save"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer-save}{}}}
\subsection{Method \code{save()}}{
Save the plot to a file. The format is determined by the file extension.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer$save(filename, width = NULL, height = NULL, dpi = 300, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{(`character(1)`)\cr
The filename to save the plot to. The file extension determines the format.}

\item{\code{width}}{(`numeric(1)`)\cr
Width of the plot in pixels (for plotly) or inches (for ggplot2).}

\item{\code{height}}{(`numeric(1)`)\cr
Height of the plot in pixels (for plotly) or inches (for ggplot2).}

\item{\code{dpi}}{(`numeric(1)`)\cr
Resolution for ggplot2 plots (ignored for plotly plots).}

\item{\code{...}}{Additional arguments passed to the underlying save function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
