% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer1D.R
\name{Visualizer1D}
\alias{Visualizer1D}
\title{Visualize Base Class}
\description{
This class is used to create 1D visualizations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x}}{(`numeric(n)`)\cr
x-values of function}

\item{\code{y}}{(`numeric(n)`)\cr
y-values of function}

\item{\code{title}}{(character(1)\cr
Title of plot}

\item{\code{lab_x}}{(character(1)\cr
Label of x-axis}

\item{\code{lab_y}}{(character(1)\cr
Label of y-axis}

\item{\code{x}}{(`numeric(m)`)\cr
x-values of extra points to plot.
Use NULL if no points should be plotted.}

\item{\code{y}}{(`numeric(m)`)\cr
y-values of extra points to plot.
Use NULL if no points should be plotted.}

\item{\code{line_col}}{(character(1)\cr
Color of plotted line}

\item{\code{line_width}}{(numeric(1)\cr
Width of plotted line}

\item{\code{line_type}}{(character(1)\cr
Type of plotted line}

\item{\code{points_col}}{(character(1)\cr
Color of plotted points}

\item{\code{points_size}}{(numeric(1)\cr
Size of plotted points}

\item{\code{points_shape}}{(integer(1)\cr
Shape of plotted points}

\item{\code{points_alpha}}{(numeric(1)\cr
Alpha blending of plotted points}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{x}}{(`numeric(n)`)\cr
x-values of function}

\item{\code{y}}{(`numeric(n)`)\cr
y-values of function}

\item{\code{x}}{(`numeric(m)`)\cr
x-values of extra points to plot.
Use NULL if no points should be plotted.}

\item{\code{y}}{(`numeric(m)`)\cr
y-values of extra points to plot.
Use NULL if no points should be plotted.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer1D-new}{\code{Visualizer1D$new()}}
\item \href{#method-Visualizer1D-plot}{\code{Visualizer1D$plot()}}
\item \href{#method-Visualizer1D-clone}{\code{Visualizer1D$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1D-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1D-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1D$new(
  fun_x,
  fun_y,
  title = NULL,
  lab_x = "x",
  lab_y = "y",
  points_x = NULL,
  points_y = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{(character(1)\cr
Title of plot}

\item{\code{lab_x}}{(character(1)\cr
Label of x-axis}

\item{\code{lab_y}}{(character(1)\cr
Label of y-axis}

\item{\code{points_x}}{(`numeric()`)\cr
x-values of extra points to plot.
Use NULL if no points should be plotted.}

\item{\code{points_y}}{(`numeric()`)\cr
y-values of extra points to plot.
Use NULL if no points should be plotted.}

\item{\code{x}}{(`numeric()`)\cr
x-values of function}

\item{\code{y}}{(`numeric()`)\cr
y-values of function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1D-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1D-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1D$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1D-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1D-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
