% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer1DObj.R
\name{Visualizer1DObj}
\alias{Visualizer1DObj}
\title{Visualize Objective}
\value{
Returns self invisibly.
}
\description{
This class is used to create visualizations of optimization traces.
}
\section{Super classes}{
\code{\link[vistool:Visualizer]{vistool::Visualizer}} -> \code{\link[vistool:Visualizer1D]{vistool::Visualizer1D}} -> \code{Visualizer1DObj}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objective}}{(`Objective`)\cr
The objective function.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer1DObj-new}{\code{Visualizer1DObj$new()}}
\item \href{#method-Visualizer1DObj-add_optimization_trace}{\code{Visualizer1DObj$add_optimization_trace()}}
\item \href{#method-Visualizer1DObj-plot}{\code{Visualizer1DObj$plot()}}
\item \href{#method-Visualizer1DObj-clone}{\code{Visualizer1DObj$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="add_points"><a href='../../vistool/html/Visualizer.html#method-Visualizer-add_points'><code>vistool::Visualizer$add_points()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="initialize_defaults"><a href='../../vistool/html/Visualizer.html#method-Visualizer-initialize_defaults'><code>vistool::Visualizer$initialize_defaults()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="save"><a href='../../vistool/html/Visualizer.html#method-Visualizer-save'><code>vistool::Visualizer$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1DObj-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1DObj-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1DObj$new(objective, xlim = NULL, n_points = 100L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(`Objective`)\cr
The objective to optimize.}

\item{\code{xlim}}{(`numeric(2)`)\cr
Limits for the x-axis. If NULL, will be determined from objective bounds.}

\item{\code{n_points}}{(`integer(1)`)\cr
The number of generated point per dimension.
Note that a grid of `npoints^2` values is generated and evaluated by `objective$eval(x)` to plot the surface.}

\item{\code{n_points}}{(`integer(1)`)\cr
The number of generated point per dimension.
Note that a grid of `npoints^2` values is generated and evaluated by `objective$eval(x)` to plot the surface.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1DObj-add_optimization_trace"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1DObj-add_optimization_trace}{}}}
\subsection{Method \code{add_optimization_trace()}}{
Add optimization trace to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1DObj$add_optimization_trace(optimizer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{(`Optimizer`)\cr
The optimizer to add to the plot.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1DObj-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1DObj-plot}{}}}
\subsection{Method \code{plot()}}{
Create and return the ggplot2 plot with optimization traces.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1DObj$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the parent plot method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer1DObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer1DObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer1DObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
