% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer2D.R
\name{Visualizer2D}
\alias{Visualizer2D}
\title{Visualize Base Class}
\description{
This class is used to create 2D visualizations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{grid}}{(`list()`)\cr
List with the `x1` and `x2` grid.}

\item{\code{zmat}}{(`matrix()`)\cr
The result of evaluation at each element of the cross product of `grid$x1` and `grid$x2`.}

\item{\code{title}}{(character(1)\cr
Label of the plot.}

\item{\code{lab_x1}}{(character(1)\cr
Label of the x1 axis.}

\item{\code{lab_x2}}{(character(1)\cr
Label of the x2 axis.}

\item{\code{z_lab}}{(character(1)\cr
Label of the z axis.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer2D-new}{\code{Visualizer2D$new()}}
\item \href{#method-Visualizer2D-plot}{\code{Visualizer2D$plot()}}
\item \href{#method-Visualizer2D-clone}{\code{Visualizer2D$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2D-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2D-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2D$new(
  fun_x1,
  fun_x2,
  title = NULL,
  lab_x1 = "x1",
  lab_x2 = "x2",
  z_lab = "z"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{(`character(1)`)\cr
Label of the plot.}

\item{\code{lab_x1}}{(`character(1)`)\cr
Label of the x1 axis.}

\item{\code{lab_x2}}{(`character(1)`)\cr
Label of the x2 axis.}

\item{\code{z_lab}}{(`character(1)`)\cr
Label of the z axis.}

\item{\code{grid}}{(`list()`)\cr
List with the `x1` and `x2` grid.}

\item{\code{zmat}}{(`matrix()`)\cr
The result of evaluation at each element of the cross product of `grid$x1` and `grid$x2`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2D-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2D-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2D$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2D-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2D-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
