% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer2D.R
\name{Visualizer2D}
\alias{Visualizer2D}
\title{Visualize Base Class}
\description{
This class is used to create 2D visualizations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fun_x1}}{(`numeric(n)`)}

\item{\code{fun_x2}}{(`numeric(n)`)}

\item{\code{fun_y}}{(`numeric(n)`)}

\item{\code{title}}{(`character(1)`)}

\item{\code{lab_x1}}{(`character(1)`)}

\item{\code{lab_x2}}{(`character(1)`)}

\item{\code{lab_y}}{(`character(1)`)}

\item{\code{points_x1}}{(`numeric()`)\cr}

\item{\code{points_x2}}{(`numeric()`)\cr}

\item{\code{points_y}}{(`numeric()`)\cr
y-values of points.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer2D-new}{\code{Visualizer2D$new()}}
\item \href{#method-Visualizer2D-plot}{\code{Visualizer2D$plot()}}
\item \href{#method-Visualizer2D-clone}{\code{Visualizer2D$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2D-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2D-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2D$new(
  fun_x1,
  fun_x2,
  fun_y,
  title = NULL,
  lab_x1 = "x1",
  lab_x2 = "x2",
  lab_y = "y"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun_x1}}{(`numeric()`)
x-values of function.}

\item{\code{fun_x2}}{(`numeric()`)
x-values of function.}

\item{\code{fun_y}}{(`numeric()`)
y-values of function.}

\item{\code{title}}{(`character(1)`)
Title of plot.}

\item{\code{lab_x1}}{(`character(1)`)
Label of x-axis.}

\item{\code{lab_x2}}{(`character(1)`)
Label of x-axis.}

\item{\code{lab_y}}{(`character(1)`)
Label of y-axis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2D-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2D-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2D$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2D-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2D-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
