% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer2DModel.R
\name{Visualizer2DModel}
\alias{Visualizer2DModel}
\title{Visualize Model}
\description{
This class is used to create 2D visualizations of learners and tasks.
}
\section{Super class}{
\code{\link[vistool:Visualizer2D]{vistool::Visualizer2D}} -> \code{Visualizer2DModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{(`mlr3::Task`)\cr
Task used to train the model.}

\item{\code{learner}}{(`mlr3::Learner`)\cr
Learner used to train the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer2DModel-new}{\code{Visualizer2DModel$new()}}
\item \href{#method-Visualizer2DModel-add_training_data}{\code{Visualizer2DModel$add_training_data()}}
\item \href{#method-Visualizer2DModel-init_layer_contour}{\code{Visualizer2DModel$init_layer_contour()}}
\item \href{#method-Visualizer2DModel-add_decision_boundary}{\code{Visualizer2DModel$add_decision_boundary()}}
\item \href{#method-Visualizer2DModel-clone}{\code{Visualizer2DModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer2D" data-id="add_layer_hessian"><a href='../../vistool/html/Visualizer2D.html#method-Visualizer2D-add_layer_hessian'><code>vistool::Visualizer2D$add_layer_hessian()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer2D" data-id="add_layer_taylor"><a href='../../vistool/html/Visualizer2D.html#method-Visualizer2D-add_layer_taylor'><code>vistool::Visualizer2D$add_layer_taylor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer2D" data-id="add_optimization_trace"><a href='../../vistool/html/Visualizer2D.html#method-Visualizer2D-add_optimization_trace'><code>vistool::Visualizer2D$add_optimization_trace()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer2D" data-id="init_layer_surface"><a href='../../vistool/html/Visualizer2D.html#method-Visualizer2D-init_layer_surface'><code>vistool::Visualizer2D$init_layer_surface()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer2D" data-id="plot"><a href='../../vistool/html/Visualizer2D.html#method-Visualizer2D-plot'><code>vistool::Visualizer2D$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer2D" data-id="set_layout"><a href='../../vistool/html/Visualizer2D.html#method-Visualizer2D-set_layout'><code>vistool::Visualizer2D$set_layout()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer2D" data-id="set_scene"><a href='../../vistool/html/Visualizer2D.html#method-Visualizer2D-set_scene'><code>vistool::Visualizer2D$set_scene()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DModel-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DModel$new(
  task,
  learner,
  x1_limits = NULL,
  x2_limits = NULL,
  padding = 0,
  n_points = 100L,
  training_points = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{([mlr3::Task])\cr
The task to train the model on.}

\item{\code{learner}}{([mlr3::Learner])\cr
The learner to train the model with.}

\item{\code{x1_limits}}{(`numeric(2)`)\cr
Limits for the first feature axis.}

\item{\code{x2_limits}}{(`numeric(2)`)\cr
Limits for the second feature axis.}

\item{\code{padding}}{(`numeric(1)`)\cr
Padding around the data limits.}

\item{\code{n_points}}{(`integer(1)`)\cr
Number of points to use for visualization.}

\item{\code{training_points}}{(`logical(1)`)\cr
Whether to show training points on the plot.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DModel-add_training_data"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DModel-add_training_data}{}}}
\subsection{Method \code{add_training_data()}}{
Adds the training data to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DModel$add_training_data(size = 2, color = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{(`numeric(1)`)\cr
Size of the points.}

\item{\code{color}}{(`character(1)`)\cr
Color of the points.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DModel-init_layer_contour"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DModel-init_layer_contour}{}}}
\subsection{Method \code{init_layer_contour()}}{
Initialize the contour layer. For ggplot2-based 2D visualizers, this method 
exists for API compatibility but doesn't need to do anything since contours 
are created by default in the plot() method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DModel$init_layer_contour(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments (ignored for ggplot2 compatibility).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DModel-add_decision_boundary"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DModel-add_decision_boundary}{}}}
\subsection{Method \code{add_decision_boundary()}}{
Adds the decision boundary to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DModel$add_decision_boundary(threshold = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{(`numeric(1)`)\cr
Threshold for the decision boundary.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
