% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer2DModel.R
\name{Visualizer2DModel}
\alias{Visualizer2DModel}
\title{Visualize Model}
\description{
This class is used to create 2D visualizations of learners and tasks.
}
\section{Super class}{
\code{\link[vistool:Visualizer2D]{vistool::Visualizer2D}} -> \code{Visualizer2DModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{(`mlr3::Task`)\cr
Task used to train the model.}

\item{\code{learner}}{(`mlr3::Learner`)\cr
Learner used to train the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer2DModel-new}{\code{Visualizer2DModel$new()}}
\item \href{#method-Visualizer2DModel-add_training_data}{\code{Visualizer2DModel$add_training_data()}}
\item \href{#method-Visualizer2DModel-add_decision_boundary}{\code{Visualizer2DModel$add_decision_boundary()}}
\item \href{#method-Visualizer2DModel-clone}{\code{Visualizer2DModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer2D" data-id="plot"><a href='../../vistool/html/Visualizer2D.html#method-Visualizer2D-plot'><code>vistool::Visualizer2D$plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DModel-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DModel$new(
  task,
  learner,
  x1_limits = NULL,
  x2_limits = NULL,
  padding = 0,
  n_points = 100L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{([mlr3::Task])\cr
The task to train the model on.}

\item{\code{learner}}{([mlr3::Learner])\cr
The learner to train the model with.}

\item{\code{x1_limits}}{(`numeric(2)`)\cr
The x1 limits.}

\item{\code{x2_limits}}{(`numeric(2)`)\cr
The x2 limits.}

\item{\code{padding}}{(`numeric(1)`)\cr
A margin that is added to x1limits and x2limits.
The x1 margin is calculated by `max(x1lmits) - min(x1limits) * padding`.}

\item{\code{n_points}}{(`integer(1)`)\cr
The number of generated point per dimension.
Note that a grid of `npoints^2` values is generated and evaluated by `objective$eval(x)` to plot the surface.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DModel-add_training_data"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DModel-add_training_data}{}}}
\subsection{Method \code{add_training_data()}}{
Adds the training data to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DModel$add_training_data(size = 5, color = "grey", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{(`numeric(1)`)\cr
Size of the points.}

\item{\code{color}}{(`character(1)`)\cr
Color of the points.}

\item{\code{...}}{(`any`)\cr
Further arguments passed to `add_trace(...)`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DModel-add_decision_boundary"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DModel-add_decision_boundary}{}}}
\subsection{Method \code{add_decision_boundary()}}{
Adds the decision boundary to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DModel$add_decision_boundary(
  threshold = 0.5,
  surfacecolor = list(c(0, 1), c("rgb(176,196,222)", "rgb(160,82,45)")),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{(`numeric(1)`)\cr
Threshold for the decision boundary.}

\item{\code{surfacecolor}}{(`list()`)\cr
The coloring of the surface.}

\item{\code{...}}{(`any`)\cr
Further arguments passed to `add_trace(...)`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
