% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer2DObj.R
\name{Visualizer2DObj}
\alias{Visualizer2DObj}
\title{Visualize Objective}
\description{
This class is used to create visualizations and animations of optimization traces.
}
\section{Super classes}{
\code{\link[vistool:Visualizer]{vistool::Visualizer}} -> \code{\link[vistool:Visualizer2D]{vistool::Visualizer2D}} -> \code{Visualizer2DObj}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objective}}{(`Objective`)\cr
The objective function.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Visualizer2DObj-new}{\code{Visualizer2DObj$new()}}
\item \href{#method-Visualizer2DObj-add_optimization_trace}{\code{Visualizer2DObj$add_optimization_trace()}}
\item \href{#method-Visualizer2DObj-plot}{\code{Visualizer2DObj$plot()}}
\item \href{#method-Visualizer2DObj-clone}{\code{Visualizer2DObj$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="add_points"><a href='../../vistool/html/Visualizer.html#method-Visualizer-add_points'><code>vistool::Visualizer$add_points()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vistool" data-topic="Visualizer" data-id="save"><a href='../../vistool/html/Visualizer.html#method-Visualizer-save'><code>vistool::Visualizer$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DObj-new"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DObj-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DObj$new(
  objective,
  x1_limits = NULL,
  x2_limits = NULL,
  padding = 0,
  n_points = 100L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(`Objective`)\cr
The objective to optimize.}

\item{\code{x1_limits}}{(`numeric(2)`)\cr
The x1 limits.}

\item{\code{x2_limits}}{(`numeric(2)`)\cr
The x2 limits.}

\item{\code{padding}}{(`numeric(1)`)\cr
A margin that is added to x1limits and x2limits.
The x1 margin is calculated by `max(x1lmits) - min(x1limits) * padding`.}

\item{\code{n_points}}{(`integer(1)`)\cr
The number of generated point per dimension.
Note that a grid of `npoints^2` values is generated and evaluated by `objective$eval(x)` to plot the surface.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DObj-add_optimization_trace"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DObj-add_optimization_trace}{}}}
\subsection{Method \code{add_optimization_trace()}}{
Add optimization trace to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DObj$add_optimization_trace(
  optimizer,
  line_color = NULL,
  line_width = 1.2,
  line_type = "solid",
  npoints = NULL,
  npmax = NULL,
  name = NULL,
  add_marker_at = 1,
  marker_size = 3,
  marker_shape = 16,
  marker_color = NULL,
  show_start_end = TRUE,
  alpha = 0.8
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{(`Optimizer`)\cr
The optimizer from which the archive is extracted and used to plot the trace.}

\item{\code{line_color}}{(`character(1)`)\cr
The color of the trace line. If NULL, a random color is generated.}

\item{\code{line_width}}{(`numeric(1)`)\cr
The width of the trace line. Default is 1.2.}

\item{\code{line_type}}{(`character(1)`)\cr
The line type for the trace. Default is "solid".}

\item{\code{npoints}}{(`integer(1)`)\cr
The number of used points from the archive. Default is NULL which means all points are used.}

\item{\code{npmax}}{(`integer(1)`)\cr
The maximum number of points to use. Default is NULL.}

\item{\code{name}}{(`character(1)`)\cr
The name of the trace in the legend. Default is NULL which creates a name from optimizer and objective IDs.}

\item{\code{add_marker_at}}{(`integer()`)\cr
Vector of iterations at which markers are added. Default is first iteration only.}

\item{\code{marker_size}}{(`numeric(1)`)\cr
Size of the markers. Default is 3.}

\item{\code{marker_shape}}{(`character(1)` or `integer(1)`)\cr
Shape of the markers. Default is "circle" (16).}

\item{\code{marker_color}}{(`character(1)`)\cr
Color of the markers. If NULL, uses line_color.}

\item{\code{show_start_end}}{(`logical(1)`)\cr
Whether to highlight start and end points differently. Default is TRUE.}

\item{\code{alpha}}{(`numeric(1)`)\cr
Transparency level for the trace. Default is 0.8.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns self invisibly for method chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DObj-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DObj-plot}{}}}
\subsection{Method \code{plot()}}{
Create and return the ggplot2 plot with optimization traces.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DObj$plot(text_size = 11, theme = "minimal")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text_size}}{(`numeric(1)`)\cr
Base text size for plot elements. Default is 11.}

\item{\code{theme}}{(`character(1)`)\cr
ggplot2 theme to use. One of "minimal", "bw", "classic", "gray", "light", "dark", "void". Default is "minimal".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Visualizer2DObj-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Visualizer2DObj-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Visualizer2DObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
